Refactor experiments to allow scenarios and reports to be parameterized by composable configurable components.
Introduce a Configurable class which represents anything that has configurable attributes and can also be an injectable attribute type. For example, a scenario has a dataset attribute which is a configurable dataset object with its trainsize, valsize and testsize arguments. This approach is applied to datasets, models, pipelines and postprocessors. All these objects are refactored to have a load or construct class for returning actual objects that are usable in the scenario (e.g. a model.construct() returns a sklearn model). All scenarios are updated accordingly.